backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
groq==0.9.0
yfinance==0.2.28
pandas==2.1.4
numpy==1.26.2
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
setuptools==69.0.3
wheel==0.42.0
